add missing workflow files
